[btree] Keep the dm_btree_info instance in guest's memory
Previously, the btree wrappers require copying the BTreeInfo object
into the guest memory on every btree function call, which is
inefficient, also makes the in-kernel struct non-reusable for
dm_btree_cursor. Now add a helper function to initialize a dm_btree_info
instance in the guest, and modify the wrapper interfaces to take the
guest pointer directly.